# Optimize `File.allDeclarationsByType`