--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -6,7 +6,8 @@
 
 ##### Enhancements
 
-* None.
+* Improve performance of `MissingDocsRule`.  
+  [Norio Nomura](https://github.com/norio-nomura)
 
 ##### Bug Fixes